refactor: separate component for design part of card dialog
